+++ x/distribution/migrations/v4/migrate_funds_test.go
@@ -37,13 +37,14 @@ func TestFundsMigration(t *testing.T) {
 	cms := integration.CreateMultiStore(keys, logger)
 	encCfg := moduletestutil.MakeTestEncodingConfig(codectestutil.CodecOptions{}, auth.AppModule{}, bank.AppModule{}, distribution.AppModule{})
 	ctx := sdk.NewContext(cms, true, logger)
-
+	addressCodec := addresscodec.NewBech32Codec(sdk.Bech32MainPrefix)
 	maccPerms := map[string][]string{
 		pooltypes.ModuleName: nil,
 		disttypes.ModuleName: {authtypes.Minter},
 	}
 
-	authority := authtypes.NewModuleAddress("gov")
+	authority, err := addressCodec.BytesToString(authtypes.NewModuleAddress("gov"))
+	require.NoError(t, err)
 
 	// create account keeper
 	accountKeeper := authkeeper.NewAccountKeeper(
@@ -51,9 +52,9 @@ func TestFundsMigration(t *testing.T) {
 		encCfg.Codec,
 		authtypes.ProtoBaseAccount,
 		maccPerms,
-		addresscodec.NewBech32Codec(sdk.Bech32MainPrefix),
+		addressCodec,
 		sdk.Bech32MainPrefix,
-		authority.String(),
+		authority,
 	)
 
 	// create bank keeper
@@ -62,7 +63,7 @@ func TestFundsMigration(t *testing.T) {
 		encCfg.Codec,
 		accountKeeper,
 		map[string]bool{},
-		authority.String(),
+		authority,
 	)
 
 	// gomock initializations
@@ -79,7 +80,7 @@ func TestFundsMigration(t *testing.T) {
 		stakingKeeper,
 		poolKeeper,
 		disttypes.ModuleName,
-		authority.String(),
+		authority,
 	)
 
 	// Set feepool
@@ -87,7 +88,7 @@ func TestFundsMigration(t *testing.T) {
 	feepool := disttypes.FeePool{
 		CommunityPool: sdk.NewDecCoinsFromCoins(poolAmount),
 	}
-	err := distrKeeper.FeePool.Set(ctx, feepool)
+	err = distrKeeper.FeePool.Set(ctx, feepool)
 	require.NoError(t, err)
 
 	distrAcc := authtypes.NewEmptyModuleAccount(disttypes.ModuleName)